Stop epoll from looping so often
Added a ready table so epoll can track which of it's child FDs
actually have an event that needs to get delivered, rather than
looping through the entire child table every time the status of a
single child changes.

This change cut about 40% off of the runtime of a small Tor
experiment.