Updates to TCP buffer size and packet headers
TCP buffer sizes are now allowed to be autotuned to a higher
value that is computed based on the rtt between the two connection
endpoints.

TCP packets now get added to the retransmit queue immediately
before they get sent to the network, rather when the packet is
initially created. We also update other TCP header values (e.g.
windows and timestamps) right before sending the packet.