Add return statement to operator overload
The method has a reference to its class as return type, but does not actually
have a return statement. This has resulted in null-dereferences and segfaults
on Kali-Linux. This fix appears to resolve #699. The number of failed tests
has gone from 18 to 20 to only 1 and the functional tests exits successfully.