Refactor autotune timing to fix sockbuf test
Rather than initializing buffer sizes when the connection is first
established, we instead initialize them when we get the first new
rtt measurement.

Most of this code is commit is just moving around code. The main
change is adding the flag tcp->autotune.didInitializeBufferSizes
flag and changing where _tcp_tuneInitialBufferSizes is first
called.