Redesign router and router queue management
The queue manager logic is now separated from the router, so that
we can support multiple queue management algorithms.

Two simple algorithms are supported now: a single packet queue,
and a queue that can grow to a static size.

Also, a queue manager based on CoDel is stubbed out but not yet
functional.