Add a .clang-format configuration
This specifies the formatting preferences for the clang-format tool,
which should pick up this configuration automatically.  I started with
clang-format's defaults and made a few tweaks in accordance with the
preferences observed in the existing code.

I don't plan or recommend mass reformatting, but this is useful
for keeping modified code tidy and formatted consistently.
A mass reformat might make sense down the line once we've had time
to get comfortable with the options, and don't have any large
pending changes (since it'd make merging a huge headache).