Adds syscall-wrapper-generation-macros
These make the definitions of the more straightforward wrappers
relatively succinct, and allow us to move the check for whether we're
interposing there. That in turn lets us get rid of some of the redundant
argument unpacking and repacking.

Also fixes a bug that I'd spuriously removed the explicit constructor
priorities in the shim and the "_real_syscall" constructor.