Skeleton udp test using Rust
CMakeRust only supports building Rust static libraries, so this uses a
small C wrapper to call the Rust code.

This version of it (calling libc's `write`) succeeds, and causes the
output to be written to the expected log file.

Calling `println!` in the Rust code when running under Shadow results in
the Rust internals calling abort. Running gdb running shadow running the
test results in a segfault when trying to load the plugin.