Beginning of UDP test
This is mostly getting in place a framework for UDP tests; particularly
for using dynamically assigned ports when not running under Shadow and
communicating them via a fifo(7). I'll flesh out some more tests in a
follow-up.

Initially I tried doing the same under Shadow as well, but it doesn't
deal well with blocking on an `open` call. I got it to sort of work if I
opened the fifo in nonblocking mode and looped, but decided it'd be
cleaner to just use hard-coded ports under Shadow for now. I expect that
the blocking `open` call won't be an issue under Phantom, so we can
perhaps switch it over there.