Create a utility Logger interface
Created a minimal Logger interface as a utility library outside of the
Shadow binary. When not overridden, Logger simply uses glib's logging
functionality.

Shadow's custom logger is renamed to ShdLogger and made to be a
specialization of the Logger interface.

This allows code that runs both inside and outside of the shadow
executable to use a single logging API.