Clarify 'node', 'cluster', 'edge' terminology
Remove references to "clusters". A cluster is now a "node in the topology".

Don't refer to hosts as nodes, since this could cause them to be confused with topology nodes.

Use "node" instead of "vertice" and "edge" instead of "link", since those are the concrete terms used in the topology.