Support for blocking syscalls
This adds support for syscalls that must block due to a descriptor
not being readable/writable, or to wait for a timeout, or both.
It uses the new descriptor listener class to listen for changes
on descriptors.

The new syscall handler model is that if it instructs the thread
to block after the first syscall, then the thread should call
the same syscall a second time once the thread is resumed. The
syscall handler will correctly determine the reason for becomming
unblocked and will return the correct code accordingly.