Schedule token bucket refill tasks only when needed
If the token bucket is full and we haven't tried to send or
receive anything recently, then there is no need to continue
scheduling and executing refill tasks. Here we keep track of
when all of the refill tasks would have executed if we would have
scheduled the tasks as usual, and make sure they are still
scheduled for those same times.

refs GH-756