Block autosquash commits
"fixup!" and "squash!" commits are generally created using the `--fixup`
and `--squash` flags to `git commit`, and are intended to be squashed
into another commit before merging, generally using the `--autosquash`
flag with `git rebase`. They are a useful alternative to
amending/rebasing commits immediately, especially for a pull request
that's in the middle of review.

Without a check like this one though, it's easy to forget to `git rebase
--autosquash` them away before merging. e.g.
3787852a8b676ddef0efce9f9380895f88e414df.