Lock block-autosquash-commits-action to a specific commit
In general it's good to lock actions to a commit or tag so that they
can't change too much underneath you. It seems particularly important in
this case since it requires passing some kind of GitHub access token. I
poked through the code and validated that it seems to just be using it
to talk to the GitHub API for its stated purpose, but I'd be nervous
about the code changing underneath us.

The latest tagged release was a bit old, so I just took the commit id of
the current master.