Process-specific descriptor tracking
This commit adds a new descriptor table module that can be used to
track a set of descriptors that have been created by each plugin
process. The related functionality has been refactored and removed
from the host module.

Because descriptors are now process-specific, the descriptor
numbers that are assigned to a descriptor will start at n=3 for
each process. So if a host is running 10 processes, it will give
out 10 n=3 descriptor numbers. This will also make it easier to
assign independent stdin and stderr files for each process, and it
will be easier to understand what state needs to be copied on a
process clone.