Send descriptor status updates immediately
Rather than scheduling a task callback to send the update to
listeners, we now do that immediately. This ensures that listeners
will get fresh status info.

In particular, this model fixes a bug where epoll was never
notified that one of its child fds status changed because the
change happened during a "syscall loop" where the plugin was
executing recv syscalls in non-blocking mode. Once all of the data
was received, the plugin called epoll_wait, but epoll still thought
the socket was readable (because the notification task was not
yet executed).

The cases where we need a task are limited to notifications that
would cause us to start running the plugin again. These cases are
now handled in the SysCallCondition module.