Fix threadptrace_terminate
This function will only ever be called when the child is already
stopped. Since it's already stopped, a signal sent via `kill` won't be
processed until if/when we resume the process.

Since we know the child is already stopped, we can deliver the signal
more directly using ptrace itself.