Use inline assembly to do real syscall
The previous approach had a couple problems that are exposed when
interposing libc's syscalls (using the recompiled libc).

* dlsym itself makes syscalls.
* libc was trying to make syscalls before our initializer ran. We might
be able to address this via lazy initialization rather than a global
constructor, but would need to be careful to avoid syscalls. (e.g.
pthread_once uses `futex`).

There might be a way to work around these problems, but I think just
using the assembly will be simpler and more robust.