threadptrace_nativeSyscall: handle variable number of ptrace stops
For reasons I haven't been able to completely nail down, when
single-stepping a syscall instruction, ptrace *sometimes* generates a
stop without actually advancing the instruction pointer.

I used to think it always did so exactly once, but it turns out it
definitely sometimes does so 0 times. We must have gotten away with this
happening before either because the rax happened to not be changed by
the instruction after the syscall, or we didn't care about the syscall
return value in the cases where it was happening.

I wish I understood better exactly what was happening here, but it's now
handled in a loop that should be fairly robust.