Force frame pointers in Rust. Useful to get sensible prof reports (#918)
This follows the existing precedent of setting -fno-omit-frame-pointer for all builds in the C code.

In principle we could omit these flags (i.e. omit the frame pointers) in release builds, but given that the performance cost is generally negligible (validated with a quick ad-hoc phold benchmark), it's worth making the release builds a bit more debuggable and profile-able.