When running from 'build/', don't use the installed shim
Previously when running shadow from the 'build' directory (for
example with './setup.py test'), shadow would use the installed
shim rather than the most-recently built shim in 'build/src/shim'.
This commit modifies how the rpath is generated and causes shadow
to use a different rpath after it is installed.