Change Thread methods to be safe, but take mutable references.
Previously some of these were marked unsafe based on being able to crash
the given thread, or being able to invalidate pointers to the thread's
memory. However, Shadow itself should be robust to the former, and
dereferencing pointers is already considered unsafe.

Conservatively marking all Thread methods that involve executing code in
the Thread as mutable for now. In particular we want to prevent letting
client code call methods on a Thread that could invalidate outstanding
references borrowed from the Thread (via the MemoryManager).