ptrace: don't double-detach
When running phold-threaded-shadow-ptrace with more hosts (e.g. 20
instead of the default 10), we'd get errors in threadptrace_doDetach. It
turns out that a host can end up calling detach on a plugin that's
already detached.

Since threadptrace_doDetach was assuming it was starting from an
attached state, things didn't go well.

We now simply silently return if already detached.