Add support for declaring options in module that owns them
The helpers currently only add to the group "experimental", though
there's no reason we couldn't also add helpers for other groups.

Doing it this way allows us to avoid piping options all the way from the
global config, through layers of code that don't understand them, to the
code that does understand and own the option.