Use sem_trywait instead of separate atomic
sem_trywait already provides an efficient way of trying to check a
semaphore without blocking. Using it instead of a separate atomic
simplifies things quite a bit.

Performance seems to be the same for a 30s phold simulation.

This also eliminates some subtle interactions between the atomic and
the semaphore.

For example:
 * Thread A: calls `wait`, hits spin threshold, blocks on `sem_wait`
 * Thread B: calls `post`
 * Thread C: calls `wait`, sees that the atomic is available and breaks
out of the loop
 * Thread A: wakes up and grabs the semaphore
 * Thread B: calls `sem_wait` expecting it not to block, but it does
block since the semaphore was taken out from under it.