Merge master, Fix inappropriate processID usage
Before merge blockchain-sim:storage_sharing, submodule should be
merged.
On merging, I found that the virtual processID is increase
between the border of the host, so that if the first virtual app.
on host 1 is 1, then the first virtual app. on host 2 is 2.
This may violate the isolated view on each host.
So I reverted the changes on processID and added bleepProcessID
for storage sharing mechanism.