The epoll tables now index their entries by (fd, object_ptr)
Rather than indexing only by the fd, the epoll descriptor 'watching'
and 'ready' tables now index their entries by a tuple of the fd and a
pointer to the object. This is similar to how Linux implements epoll,
and will give us more expected behaviour. It is also on the path to
adding support for watching Rust descriptor objects.