Fix nasty epoll bug causing infinite loops
The bug occured when an epoll is watching another epoll, which
is watching a socket:
epoll1->epoll2->socket

In this case, when the socket was done and the epoll_ctl syscall
deleted the socket from epoll2, epoll2 did not update its own
status to mark itself as no longer readable so that epoll1 would
know that its children no longer had events. This caused an
infinite loop when running tgen, where epoll1 said that epoll2
was ready, but epoll2 said it had no events.

This commit fixes it so whenever epoll_ctl changes the status of
an epoll, it updates its own status (which then propagates to any
other epolls watching it).