Reorganize how we close descriptor objects
When calling process_deregisterCompatDescriptor(), it now returns an
owned reference to a Box of the descriptor. This means that the
calling code now owns the descriptor object, and we can implement
functions that drop the descriptor. For example, we add the close()
method which takes the argument 'self' and will drop the descriptor.