thread_ptrace: Receive and process PTRACE_EVENT_EXIT
PTRACE_EVENT_EXIT is received a bit more consistently and reliably than
the "normal" WIFEXITED ptrace exit stops.

In particular, from ptrace(2): If  a  thread  group  leader  is  traced
and  exits  by   calling  _exit(2),  a PTRACE_EVENT_EXIT  stop  will
happen  for  it  (if requested), but the subsequent WIFEXITED
notification will not be delivered until all  other  threads  exit.

This led to thread_ptrace waiting forever for the next ptrace-stop in
the thread group leader when it happened.