Gracefully handle whole-process exits
When a process exits via `exit_group` or is killed by a signal, all of
its threads are also killed. The other thread deaths happen sometime
"soon" though; they're not guaranteed to already be dead the next time
we interact with them. If we attempt to go on as normal, then any call
to e.g. `ptrace` or `waitpid` must be prepared for the thread to no
longer exist.

Instead, once we've detected that the whole process is going to be
killed, we mark the process as "exiting". In that state the process
won't try to run any more threads; it'll just clean up and finish
exiting.

Similarly, the threads no longer provide a `terminate` method, since
there actually is no way to force termination of only a single thread.
Process termination via signal is moved into the `Process` object.