Make shadow logger registration thread-safe
Since workers concurrently create ForkProxy threads, we need to be able
to register those threads with the logger in a thread-safe way. This
adds a GRWLock to protect the shadow logger's hash table of
thread-specific data.

I considered doing away with the hash table entirely and using
thread-local storage, but then we wouldn't have a way to flush all of
the thread buffers at shutdown.

Fixes #1218