Decouple # of worker threads from max concurrency
1. Adds a WorkerPool with a callback-style interface instead of having a
   series of coordinated barriers.

2. Adds a --max-concurrency option that allows us to restrict the number
   of workers that can run concurrently. The WorkerPool also implements
   "worker stealing" to try to keep all --max-concurrency CPUs utilized.
   e.g. setting the # of workers to the # of hosts, max concurrency to
   the number of CPUs, and using the "host" scheduler is somewhat analagous
   to using the work-stealing scheduler.

These changes are intended primarily for the dev/phantom branch, since
ptrace has better performance with a smaller number of tracees per
thread. We're including it here in main for apples-to-apples performance
comparison between the shadow 'classic' and 'phantom' architectures.

Originally I tried to keep 1. as a separate commit from 2., but after a
series of fixup commits 1. and 2. got too tangled together.