Remove inter-round collect info step in scheduler
We stop waking threads up in between rounds in order to compute
the minimum next event time. Instead we keep track of the min time
during the execution of the round.

When using the work-stealing scheduler, we manage spinning on the
isStealable initialization flag without doing a linear scan over
the threads in between rounds to reset the flag.

closes #1249