Add process_readPtr and process_writePtr
These APIs are potentially more efficient than
process_getReadablePluginPtr and process_getWriteablePluginPtr when the
data must be copied to or from some other Shadow-owned buffer (such as a
packet). The latter can fall back to thread_* when the MemoryManager is
disabled or doesn't handle the given memory region.

For thread_ptrace, thread_getXPluginPtr involves an extra allocation and
copy to a buffer owned by ThreadPtrace, while thread_readPtr and
thread_writePtr doesn't.

These APIs also lend themselves to safe, clean Rust APIs. We should be
able to build generically typed interfaces for reading and writing
primitives, without having to deal with the lifetime issues we'd have to
contend with if returning some kind of borrowed pointer.