Add worker_readPtr and worker_writePtr
To get the benefits of process_readPtr and process_writePtr, we need to
pass virtual pointers deeper into the code. Having to also pass Process
and Thread pointers in all such cases would be cumbersome. Instead we
can extend the Worker to track the active thread and provide succinct
wrappers for these APIs.