Do all pointer access through Process
After thinking about it some more, it'd be better to not need to hold a
mutable reference to the whole Worker when we need a mutable reference
to a Process's memory. So now we use the Worker to get the active
process if needed, and then use that to do memory access.

I also changed the Process memory methods to *not* take a thread, and
instead implicitly use the active thread. The fact that it needs the
thread at all is an implementation detail that can perhaps go away in
the future.

I considered creating a SyscallManager proxy object and getting the
Process proxy object from that in the Rust code instead of from the
Worker. On reflection though, I think it might be better to eventually
remove the Process and Thread pointers from the SyscallManager and
*always* get it from the Worker.

We might even be able to make the Worker be the sole owner of the Hosts
and related objects, allowing us to store them in RefCells without Arcs
or Mutexes. For now I designed the Worker borrow APIs to take a callback
rather than returning a reference to a borrowable object, or a borrowed
object. This encapsulates the details of how to borrow the object as an
implementation detail of the Worker.