Allow NULL pointer with 0-size pointer accesses
Pointer APIs that return a slice now return an empty slice when given
size is zero, even if given a NULL pointer.

Pointer APIs that read or write a NULL pointer succeed when size is
zero, even if given a NULL pointer.

This allows syscall handlers to be lenient when passed a NULL pointer of
zero-size without having to consider this case explicitly. e.g.
experimentally in Linux `read` and `write` calls with a NULL pointer and
size 0 succeed. They now succeed in Shadow as well.