Process APIs: Check pointer alignment and help caller choose API
I ended up removing entirely the APIs that get a
readable/writable/mutable pointer to a single value. I think in any such
case it'll be preferable to use the read_ptr_* and write_ptr_* APIs,
since the size of a copy is generally trivial. We can add these back in
if we need them, but for now it helps reduce the API surface a bit.