syscallhandler_shadow_hostname_to_addr_ipv4: null-terminate name
The previous call to process_getReadablePtr wasn't including the
null-byte in the read. This happens to work in the MemoryMapper path,
since the NULL byte will still be there, but in the case where we do a
remote-read this could result in garbage at the end.

Switch to process_getReadableString, which ensures the returned  string
is null-terminated.