clone-linux: use atomic operations instead of volatile
While volatile seems to work in practice, it's not really recommended
for this purpose (or I think for any purpose in modern C). Replace them
with atomic operations, which are a supported mechanism for cross-thread
synchronization.

Made this change while debugging #1252. I don't think it turned out to
be an observed problem in practice, but ought to be fixed anyway.