Reorganize MemoryManager
The main thing I wanted to solve here is that the MemoryReader
and MemoryWriter objects were stateful in surprising ways. e.g. calling
`writer.as_mut_uninit` would permanently set that writer to an
"uninitialized" reference. Later calling `writer.as_mut` on the same
object would still give the uninitialized buffer.

To solve this, "reader" and "writer" objects are replaced with
more-focused smart-pointer objects (ProcessMemoryRef and
ProcessMemoryRefMut); the caller decides the details of what they'll
point to at time of object creation.

While temporarily disabling the MemoryMapper for debugging this change,
ran across a bug where an uninitialized buffer was being written back
into the plugin after a failed syscall.

* There are now APIs for freeing a mutable reference instead of writing
it back, which should be used if the reference was never actually
overwritten (e.g. because an error was encountered in the syscall
handling).

* For the C syscall handlers this is now automatically called in
syscallhandler_make_syscall for any failed syscall.

* In debug builds, an "uninitialized" buffer is now *always* filled with
garbage data, even if the region is mapped. This is to prevent latent
bugs from being hidden by such buffers *usually* being initialized in
practice.

Some other bug fixes and cleanups along the way:

* Split the MemoryCopier and MemoryMapper into separate modules for
better encapsulation.

* Using enums instead of trait objects, which lets the MemoryCopier and
MemoryMapper objects have more-focused APIs, and makes it easier to
follow what the code is doing.

* Marked more functions as `unsafe`, and added more `Safety` annotations
for how to call unsafe functions safely, and for justifying usage of
`unsafe` blocks.