Worker: make Rust Worker the canonical Worker object
This is a step towards moving the object graph into Rust, so that we can
sort out ownership / sharing issues.

The Rust Worker object is currently just a thin wrapper around the
legacy C Worker object (now CWorker). It should be straightforward to
start moving more members out of the CWorker directly into the Worker,
but I'd like to prioritize getting all of the Shadow objects into at
least this state before converting any one object completely.

* The pieces of the Worker that are accessed from outside the Worker are
moved into parallel arrays in the WorkerPool. This allows the Worker
itself to be stored in a thread-local without any thread-safe wrappers,
for cheaper access.

* I removed support for 0-Workers, since it was already planned, and
it'd have further complicated the previous change.