Remove C usage of getActiveHost, getActiveProcess, getActiveThread
This is conceptually a follow-up to #1386, but doesn't actually depend
on it. Once both this and #1386 are in, we can remove getActiveHost,
getActiveProcess, and getActiveThread entirely. This is helpful for
porting to Rust since we don't have or want those to be globals there.

In C code it suffices to pass around the "deepest" current context
object, since they all have "up" pointers. e.g. given a Thread we can
get the corresponding Process and Host. We might eventually want to
replace these parameters with their Context counterparts in #1386, but
this seems like a useful intermediate step.