test_clone: Fix and extend clone_child_exits_after_leader
Allowing `main` to exit was causing the whole process to exit (via
exit_group) instead of just exiting the thread-group-leader.

This resulted in no-longer exercising the edge case where the
MemoryCopier tries to use the task id of a dead thread (the thread group
leader).

Also extended this test to more explicitly test this, by having another
thread validate that Shadow was able to clear the ctid after one of the
other threads exited (both after the thread group leader exited).