Worker: simplify interfaces
Add internal helper for borrowing the WORKER.

Don't expose references to the Worker to C code; only operations.  The
idea behind the references were partially to guarantee atomicity across
multiple operations, and partly to save having to borrow the Worker
multiple times.

Atomicity isn't much of an issue since the Worker is thread-local. In
principle a function call could unexpectedly mutate the Worker, but it's
unlikely in practice.

For performance, the extra heap allocation and deallocation (each of
which require an atomic operation) needed to wrap the reference object
likely outweighs the cost of doing multiple RefCell borrows (which
require no atomic operations).