Handle unaligned pointers from managed processes gracefully
In Rust code, dereferencing an unaligned pointer is undefined behavior.
For that reason, we were returning errors if a managed plugin passed an
unaligned pointer to a syscall.

In Linux, though, unaligned pointers are permitted and handled
gracefully.

With this change, we now also handle unaligned pointers transparently in
the MemoryManager. The important thing is that we don't try turning them
into a reference in Rust, which we should be able to avoid by falling
back to copying data to and from such pointers (as opaque data).

I think the main way we could run into trouble is if in, say, a
syscallhandler, we ask the MemoryManager for a reference to a u8 pointer
but then cast it to some other type without checking alignment. This can
and should be avoided by asking the MemoryManager for the desired type
in the first place.

Right now, C syscallhandlers always ask for u8 pointers since we don't
have an easy way to convey the type to Rust. This should mostly work out
ok in practice, since in x86 unaligned pointer access is legal for most
instructions (bug e.g. not vectorized ones). (This change doesn't make
things any worse, since we weren't catching when this happens before,
either).