Managed -> Virtual Process in most places
For the most part, the user doesn't need to know that processes in a
simulation are implemented using native processes on the host machine.
In those places, the more abstract term "virtual process" makes more
sense.

When referring specifically to the real process running on the host OS,
"managed process" is probably still the best term. This comes up e.g. in
developer/debugging docs, and in some experimental options.