Report epoll events in a deterministic order
When a managed process attempts to retrieve epoll events for fds
that it is waiting on, we need to return the fds to the process
in an order that is deterministic across simulation runs.

Here we add a mechanism to sort the watch objects so that we
report the events in a deterministic order.