Add an option to use a socket for ipc
This was motivated by seccomp use-cases of being able to send a
file-descriptor over the socket. As it turns out, I'm not sure we need
that capability after all. It is a bit easier to whitelist the
corresponding IPC in the seccomp filter though, and a bit simpler and
more flexible in general.

It *should* have similar performance to the semaphore approach when
we're not spinning, but I've left it in as an option so we can verify.