Extract syscalls that we preload to a new lib
This allows us to selectively disable the preloading of syscalls,
meaning the interposition would be done using only the secondary
seccomp or ptrace methods instead.

We normally want to keep syscall preloading enabled when it is
combined with seccomp, because it means we don't have to go through
the seccomp path and therefore we can save some overhead
(mode transitions, running the seccomp filter, calling the callback).

For ptrace, even if we preload we end up just making a syscall and
letting ptrace handle it anyway, so preloading may add some
insignificant amount of overhead.