Adds a script and new definitions for preloaded syscall wrappers
The script fetches the lastest syscall def file from the Linux
source tree and produces a list of functions that we can preload
to intercept the syscall wrappers in libc. The goal is to intercept
all wrapper functions even if Shadow does not yet handle them all.

This commit effectively removes two functions from the list of those
that were previously defined in our preload lib:
  - prlimit()
    According to the man page:
    "The name of the glibc wrapper function is prlimit(); the underlying
    system call is prlimit64()."
  - mmap2()
    According to the man page:
    "This system call does not exist on x86-64."