Don't panic on non-unicode env vars (#1664)
In particular this avoids a panic when using the HEAPPROFILE env
variable to enable gperftools heap profiling
https://gperftools.github.io/gperftools/heapprofile.html.

I suspect it has something to do with "Due to a hack we make to work
around a possible gcc bug, your profiles may end up named strangely if
the first character of your HEAPPROFILE variable has ascii value greater
than 127". It looks like the referenced hack is mangling the env var in
memory to something non-unicode.