Return replaced desc from 'descriptortable_set()'
In general, the calling function would want to close the descriptor
after it's removed rather than immediately dropping/freeing it, so
returning the descriptor allows the calling function to do that. But
right now, we know that 'descriptortable_set()' is only called for
fds 0, 1, and 2, and that they won't have previously been used, so
we can assume that one isn't returned.