Renamed 'FileStatus' to 'FileState' and 'FileFlags' to 'FileStatus'
The 'FileFlags' are called 'status flags' in Linux so it was renamed
to 'FileStatus'. We needed to rename the existing `FileStatus' flags
which represent the readable/writable/closed/etc status of the file,
so it is now 'FileState'. Other related names such as
'StatusEventSource' have also been renamed to match.