fake sigaltstack
This prevents managed calls to sigaltstack from interfering with the
shim's own configured sigaltstack.

The current fakery should be good enough until if and when a managed
program's signal handler is invoked. preload mode is already broken in
that case, though.

Once signal handlers in preload mode are working, we probably ought to
emulate this more faithfuly, since it may be difficult to debug if the
program goes wrong because *our* altstack wasn't big enough, or it
somehow detected it wasn't running on the configured stack etc.