New option to turn on crypto preloading
The option is false by default. If set to true, we will add the
openssl crypto preload lib to the LD_PRELOAD path for all nodes
running in Shadow to skip some encrypt and decrypt operations.
This may speed up simulations, particularly on older CPUs that
do not support AES-NI optimizations.

Turning on crypto preloading should be done with care, since it
may have unintended side effects for the applications running in
Shadow.