shim_enableInterposition + shim_disableInterposition -> shim_swapAllowNativeSyscalls (#1852)
The old API assumed a "stack" of disabling interposition, where we
want to ensure interposition is disabled while beginning some operation,
and then restore the current state afterwards. It doesn't support the
inverse - ensuring interposition is *enabled* for some operation.

The new api allows pushing and popping the current state in a more
general way. In particular this will allow us to re-enable interposition
when invoking a managed process's signal handler.

It'll also allow us to ensure interposition is enabled when making
"shadow syscalls" in preload mode, from a context where interposition
was disabled higher in the stack.