Intercept EVP_EncryptUpdate calls not made from libssl
We keep a small cache of callers into EVP_EncryptUpdate so we don't
have to perform expensive string operations on every call. For now
we just use a small cache of pointers and a dead simple caching
policy; this works fine in Tor, where testing reveals that there is
a small number of callers of EVP_EncryptUpdate from tor/libssl.so.