test_sendto_recvfrom: tolerate buffered reads (#1894)
This test was occasionally failing when run under Linux; e.g.:
https://github.com/shadow/shadow/runs/5100489526?check_suite_focus=true#step:8:269

It seems possible that Linux might have separate buffers for the read
and write end of the sockets, and maybe even run TCP between them, which
could result in the reader blocking before all the written data is
available to it. This change prevents the test from failing in that case
and retries until all data has been read or the test times out.