Restructured the 'PosixFile' and 'Descriptor' types
The `PosixFile` has been split into the `OpenFile` and `GenericFile` types. The
`OpenFile` is similar to a posix file description (linux `struct file`), and
the `GenericFile` is simply an enum that wraps the specific file types like
`PipeFile`, `SocketFile`, etc.

A posix file description represents an open file. You should be able to hold a
reference to a file description to keep the file open. For example in linux, a
syscall such as `read()` holds a reference to the file description so that the
file object will not close during the lifetime of the syscall, even if that
syscall blocks. With shadow's current design where the open count is stored in
the `Descriptor` object itself, it's not possible to hold a reference to the
`PosixFile` object and ensure that the file doesn't close. This means that it's
not possible to keep a file open during the lifetime of a syscall, which can
lead to that syscall blocking forever with no way to interrupt it.

This commit is mostly a structure change, but there is one significant
difference in behaviour. A syscall condition object for a blocked syscall
handler will keep the file it's waiting on open. This means that even if all of
the file handles (fds) for that file are closed in other threads, the file
object itself won't be closed until after the blocked syscall is complete. This
behaviour is how linux operates.  But there are still differences between
shadow's and linux's behaviour. For example shadow re-runs syscall handlers
rather than resuming them. A second invocation of the syscall handler would be
unable to find the file handle (fd) in the file descriptor table, so would
return `EBADF`, whereas linux would resume and complete the syscall.