Allow "null" values for some cli options
There is currently no way to set a "null" state for cli options with clap, even
when that same option supports "null" values in the configuration file. This
change allows you to specify "null" for any cli option that uses the
`NullableOption` wrapper in the configuration code.

In the shadow configuration, "null" is generally used to mean "disabled". For
example, the "template_directory" (`Option<String>`) option is disabled by
setting it to "null" (and is set to "null" by default since shadow doesn't use
a template directory by default), which corresponds to a `None` value in the
rust code.